+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
  * Added method `\TgBotLib\Bot > addStickerToSet()` to add a new sticker to a set created by the bot. 
    See [addStickerToSet](https://core.telegram.org/bots/api#addstickertoset) for more information.
  * Added abstract class `\TgBotLib\Abstracts > StickerFormat` to represent a sticker format ("`static`", "`animated`" or "`video`")
+ * Added method `\TgBotLib\Bot > uploadStickerFile()` to upload a sticker file with a sticker for later use in `createNewStickerSet` and `addStickerToSet` methods (can be used multiple times). 
+   See [uploadStickerFile](https://core.telegram.org/bots/api#uploadstickerfile) for more information.
 
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`